Crate libdonet

DONET SOFTWARE

Copyright (c) 2024, Donet Authors.

This program is free software; you can redistribute it and/or modify it under the terms of the GNU Affero General Public License version 3. You should have received a copy of this license along with this source code in a file named “LICENSE.”

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.

You should have received a copy of the GNU Affero General Public License along with this program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

libdonet

Provides the necessary utilities and definitions for using the Donet networking protocol.

These utilities include a lexer, parser, and high-level representation of the parsed DC file, as well as creating datagrams, iterating through datagrams, and the definition of every message type in the Donet protocol.

Getting Started

The recommended way to get started is to enable all features. Do this by enabling the full feature flag:

libdonet = { version = "0.1.0", features = ["full"] }

Feature Flags

The crate provides a set of feature flags to reduce the amount of compiled code. It is possible to just enable certain features over others. Below is a list of the available feature flags.

full: Enables all feature flags available for libdonet.

datagram: Includes Datagram / Datagram Iterator source for writing network packets.

dcfile: Includes the DC file lexer, parser, and DC element structures.

Modules

byte_order

Utils for swapping little-endian bytes to the compiling processor’s native endianness (byte order).

datagram

dcarray

dcfield

dcfile

dckeyword

dclass

dclexer

dcnumeric

dcparser

dcstruct

dctype

globals

Includes definitions of type aliases for Donet concepts, and the full definition of the network protocol message types.

Functions

read_dc_files

Easy to use interface for the DC file parser. Handles reading the DC files, instantiating the lexer and parser, and either returns the DCFile object or a Parse/File error.