Struct libdonet::dcfield::DCField

pub struct DCField { /* private fields */ }

A field of a Distributed Class. The DCField struct is a base for struct and dclass fields. In the DC language, there are three types of field declarations, which are: plain fields, atomic, and molecular.