Enum libdonet::globals::Protocol

#[repr(u16)]
pub enum Protocol {

    ClientHello = 1,
    ClientHelloResp = 2,
    ClientDisconnect = 3,
    ClientEject = 4,
    ClientHeartbeat = 5,
    ClientObjectSetField = 120,
    ClientObjectSetFields = 121,
    ClientObjectLeaving = 132,
    ClientObjectLeavingOwner = 161,
    ClientEnterObjectRequired = 142,
    ClientEnterObjectRequiredOther = 143,
    ClientEnterObjectRequiredOwner = 172,
    ClientEnterObjectRequiredOwnerOther = 173,
    ClientDoneInterestResp = 204,
    ClientAddInterest = 200,
    ClientAddInterestMultiple = 201,
    ClientRemoveInterest = 203,
    ClientObjectLocation = 140,
    CASetState = 1_000,
    CASetClientID = 1_001,
    CASendDatagram = 1_002,
    CAEject = 1_004,
    CADrop = 1_005,
    CAGetNetworkAddress = 1_006,
    CAGetNetworkAddressResp = 1_007,
    CADeclareObject = 1_010,
    CAUndeclareObject = 1_011,
    CAAddSessionObject = 1_012,
    CARemoveSessionObject = 1_013,
    CASetFieldsSendable = 1_014,
    CAOpenChannel = 1_100,
    CACloseChannel = 1_101,
    CAAddPostRemove = 1_110,
    CAClearPostRemoves = 1_111,
    CAAddInterest = 1_200,
    CAAddInterestMultiple = 1_201,
    CARemoveInterest = 1_203,
    SSCreateObjectWithRequired = 2_000,
    SSCreateObjectWithRequiredOther = 2_001,
    SSDeleteAIObjects = 2_009,
    SSObjectGetField = 2_010,
    SSObjectGetFieldResp = 2_011,
    SSObjectGetFields = 2_012,
    SSObjectGetFieldsResp = 2_013,
    SSObjectGetAll = 2_014,
    SSObjectGetAllResp = 2_015,
    SSObjectSetField = 2_020,
    SSObjectSetFields = 2_021,
    SSObjectDeleteFieldRAM = 2_030,
    SSObjectDeleteFieldsRAM = 2_031,
    SSObjectDeleteRAM = 2_032,
    SSObjectSetLocation = 2_040,
    SSObjectChangingLocation = 2_041,
    SSObjectEnterLocationWithRequired = 2_042,
    SSObjectEnterLocationWithRequiredOther = 2_043,
    SSObjectGetLocation = 2_044,
    SSObjectGetLocationResp = 2_045,
    SSObjectSetAI = 2_050,
    SSObjectChangingAI = 2_051,
    SSObjectEnterAIWithRequired = 2_052,
    SSObjectEnterAIWithRequiredOther = 2_053,
    SSObjectGetAI = 2_054,
    SSObjectGetAIResp = 2_055,
    SSObjectSetOwner = 2_060,
    SSObjectChangingOwner = 2_061,
    SSObjectEnterOwnerWithRequired = 2_062,
    SSObjectEnterOwnerWithRequiredOther = 2_063,
    SSObjectGetOwner = 2_064,
    SSObjectGetOwnerResp = 2_065,
    SSObjectGetZoneObjects = 2_100,
    SSObjectGetZonesObjects = 2_102,
    SSObjectGetChildren = 2_104,
    SSObjectGetZoneCount = 2_110,
    SSObjectGetZoneCountResp = 2_111,
    SSObjectGetZonesCount = 2_112,
    SSObjectGetZonesCountResp = 2_113,
    SSObjectGetChildCount = 2_114,
    SSObjectGetChildCountResp = 2_115,
    SSObjectDeleteZone = 2_120,
    SSObjectDeleteZones = 2_122,
    SSObjectDeleteChildren = 2_124,
    DBSSObjectActivateWithDefaults = 2_200,
    DBSSObjectActivateWithDefaultsOther = 2_201,
    DBSSObjectGetActivated = 2_207,
    DBSSObjectGetActivatedResp = 2_208,
    DBSSObjectDeleteFieldDisk = 2_230,
    DBSSObjectDeleteFieldsDisk = 2_231,
    DBSSObjectDeleteDisk = 2_232,
    DBCreateObject = 3_000,
    DBCreateObjectResp = 3_001,
    DBObjectGetField = 3_010,
    DBObjectGetFieldResp = 3_011,
    DBObjectGetFields = 3_012,
    DBObjectGetFieldsResp = 3_013,
    DBObjectGetAll = 3_014,
    DBObjectGetAllResp = 3_015,
    DBObjectSetField = 3_020,
    DBObjectSetFields = 3_021,
    DBObjectSetFieldIfEquals = 3_022,
    DBObjectSetFieldIfEqualsResp = 3_023,
    DBObjectSetFieldsIfEquals = 3_024,
    DBObjectSetFieldsIfEqualsResp = 3_025,
    DBObjectSetFieldIfEmpty = 3_026,
    DBObjectSetFieldIfEmptyResp = 3_027,
    DBObjectDeleteField = 3_030,
    DBObjectDeleteFields = 3_031,
    DBObjectDelete = 3_032,
    MDAddChannel = 9_000,
    MDRemoveChannel = 9_001,
    MDAddRange = 9_002,
    MDRemoveRange = 9_003,
    MDAddPostRemove = 9_010,
    MDClearPostRemoves = 9_011,

}

Enumerator for every message type in the Donet network protocol.

The Donet network protocol is made up of two separate, yet similar protocols: The Client and the Internal protocol.

Every message type defined has a certain 16-bit number identifier. Message type number IDs are organized in ranges, each range designated to a certain Donet service.

Client Protocol

All message types in the client protocol are under the <1000 range.

0-99: Basic client messages for the initial handshake with the Client Agent and maintaining the connection.

100-199: Distributed Object generates and field updates.

200-999: Client Interest controls.

Internal Protocol

1000-1999 - Client Agent

1000-1099 - Client State and Session controls

1100-1199 - Channel Controls

1200-1999 - Interest Controls

2000-2999 - State Server / DBSS

2000-2009 - Object Create / Delete

2010-2039 - Object Fields

2040-2049 - Object Location

2050-2059 - Object Designated AI

2060-2099 - Object Ownership

2100-2199 - Zone Controls

Database State Server

2200-2229 - DBSS Object Activation

2230-2239 - Object Data on Disk

3000-3999 - Database Server

3000-3009 - Object Create

3010-3019 - Object Get Fields

3020-3029 - Object Set Fields

3030-3039 - Object Delete Fields

9000-9999 - Message Director

9000-9009 - Channels and Channel Ranges

9010-9019 - Post Removes

Variants

ClientHello = 1

args(dc_hash: u32, version: &str)

This is the first message a client may send. The dc_hash is a 32-bit hash value calculated from all fields/classes listed in the client’s DC file. The version is an app/game-specific string that developers should change whenever they release a new client build. Both values are compared to the Client Agent’s DC file hash and configured version string to ensure that the client is fully up-to-date. If the client is not up-to-date, it will be disconnected with a ClientEject. If the client is up-to-date, the gameserver will send a ClientHelloResp to inform the client that it may proceed with its normal logic flow.

Excerpt taken from the Astron project, licensed under the BSD-3-Clause license.

Copyright © 2013 Sam “CFSworks” Edwards
Copyright © 2013 Kevin “Kestred” Stenerson

ClientHelloResp = 2

This is sent by the Client Agent to the client when the client’s ClientHello is accepted. This message contains no arguments.

Excerpt taken from the Astron project, licensed under the BSD-3-Clause license.

Copyright © 2013 Sam “CFSworks” Edwards
Copyright © 2013 Kevin “Kestred” Stenerson

ClientDisconnect = 3

Sent by the client when it’s closing the connection. This message contains no arguments.

ClientEject = 4

args(error_code: u16, reason: &str)

This is sent by the Client Agent to the client when the client is being disconnected. The error_code and reason arguments provide some explanation as to why the client is being dropped from the game.

Excerpt taken from the Astron project, licensed under the BSD-3-Clause license.

Copyright © 2013 Sam “CFSworks” Edwards
Copyright © 2013 Kevin “Kestred” Stenerson

ClientHeartbeat = 5

The client should send this message on a regular interval. If the Client Agent does not receive a ClientHeartbeat for a certain (configurable) amount of time, it will assume that the client has crashed and disconnect the client. This message contains no arguments.

Excerpt taken from the Astron project, licensed under the BSD-3-Clause license.

Copyright © 2013 Sam “CFSworks” Edwards
Copyright © 2013 Kevin “Kestred” Stenerson

ClientObjectSetField = 120

ClientObjectSetFields = 121

ClientObjectLeaving = 132

ClientObjectLeavingOwner = 161

ClientEnterObjectRequired = 142

ClientEnterObjectRequiredOther = 143

ClientEnterObjectRequiredOwner = 172

ClientEnterObjectRequiredOwnerOther = 173

ClientDoneInterestResp = 204

ClientAddInterest = 200

ClientAddInterestMultiple = 201

ClientRemoveInterest = 203

ClientObjectLocation = 140

CASetState = 1_000

CASetClientID = 1_001

CASendDatagram = 1_002

CAEject = 1_004

CADrop = 1_005

CAGetNetworkAddress = 1_006

CAGetNetworkAddressResp = 1_007

CADeclareObject = 1_010

CAUndeclareObject = 1_011

CAAddSessionObject = 1_012

CARemoveSessionObject = 1_013

CASetFieldsSendable = 1_014

CAOpenChannel = 1_100

CACloseChannel = 1_101

CAAddPostRemove = 1_110

CAClearPostRemoves = 1_111

CAAddInterest = 1_200

CAAddInterestMultiple = 1_201

CARemoveInterest = 1_203

SSCreateObjectWithRequired = 2_000

SSCreateObjectWithRequiredOther = 2_001

SSDeleteAIObjects = 2_009

SSObjectGetField = 2_010

SSObjectGetFieldResp = 2_011

SSObjectGetFields = 2_012

SSObjectGetFieldsResp = 2_013

SSObjectGetAll = 2_014

SSObjectGetAllResp = 2_015

SSObjectSetField = 2_020

SSObjectSetFields = 2_021

SSObjectDeleteFieldRAM = 2_030

SSObjectDeleteFieldsRAM = 2_031

SSObjectDeleteRAM = 2_032

SSObjectSetLocation = 2_040

SSObjectChangingLocation = 2_041

SSObjectEnterLocationWithRequired = 2_042

SSObjectEnterLocationWithRequiredOther = 2_043

SSObjectGetLocation = 2_044

SSObjectGetLocationResp = 2_045

SSObjectSetAI = 2_050

SSObjectChangingAI = 2_051

SSObjectEnterAIWithRequired = 2_052

SSObjectEnterAIWithRequiredOther = 2_053

SSObjectGetAI = 2_054

SSObjectGetAIResp = 2_055

SSObjectSetOwner = 2_060

SSObjectChangingOwner = 2_061

SSObjectEnterOwnerWithRequired = 2_062

SSObjectEnterOwnerWithRequiredOther = 2_063

SSObjectGetOwner = 2_064

SSObjectGetOwnerResp = 2_065

SSObjectGetZoneObjects = 2_100

SSObjectGetZonesObjects = 2_102

SSObjectGetChildren = 2_104

SSObjectGetZoneCount = 2_110

SSObjectGetZoneCountResp = 2_111

SSObjectGetZonesCount = 2_112

SSObjectGetZonesCountResp = 2_113

SSObjectGetChildCount = 2_114

SSObjectGetChildCountResp = 2_115

SSObjectDeleteZone = 2_120

SSObjectDeleteZones = 2_122

SSObjectDeleteChildren = 2_124

DBSSObjectActivateWithDefaults = 2_200

DBSSObjectActivateWithDefaultsOther = 2_201

DBSSObjectGetActivated = 2_207

DBSSObjectGetActivatedResp = 2_208

DBSSObjectDeleteFieldDisk = 2_230

DBSSObjectDeleteFieldsDisk = 2_231

DBSSObjectDeleteDisk = 2_232

DBCreateObject = 3_000

DBCreateObjectResp = 3_001

DBObjectGetField = 3_010

DBObjectGetFieldResp = 3_011

DBObjectGetFields = 3_012

DBObjectGetFieldsResp = 3_013

DBObjectGetAll = 3_014

DBObjectGetAllResp = 3_015

DBObjectSetField = 3_020

DBObjectSetFields = 3_021

DBObjectSetFieldIfEquals = 3_022

DBObjectSetFieldIfEqualsResp = 3_023

DBObjectSetFieldsIfEquals = 3_024

DBObjectSetFieldsIfEqualsResp = 3_025

DBObjectSetFieldIfEmpty = 3_026

DBObjectSetFieldIfEmptyResp = 3_027

DBObjectDeleteField = 3_030

DBObjectDeleteFields = 3_031

DBObjectDelete = 3_032

MDAddChannel = 9_000

MDRemoveChannel = 9_001

MDAddRange = 9_002

MDRemoveRange = 9_003

MDAddPostRemove = 9_010

MDClearPostRemoves = 9_011